microservices/minilm_tokenizer/requirements.txt
fastapi
pydantic
sentence-transformers
requests
python-dotenv
uvicorn
